Refactor: atomizing and blending the respective component folder to the page
