+++ tsconfig.json
@@ -16,7 +16,7 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/pages/poker-creation/index.tsx", "src/pages/testPage/index.jsx", "src/components/table-display.jsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/pages/poker-creation/index.tsx", "src/pages/testPage/index.jsx", "src/pages/poker-table/[id]/components/table-display.jsx"],
   "exclude": ["node_modules"]
 }
